Optimized mesh generation, Added features for Vis generation, Visual enhancements
Mesh Manager enables to combine multiple meshes together, store them in a optimized way and offer methods for searching specific individual meshes in the combined mesh. Implemented an Offset for the Axis and Tick generation, adapted the grid algorithm for vis container with offset on different axes. Scaled the bars and text bigger in the visualizations.